src/database/db.py
import mysql.connector
from mysql.connector import Error
from config.database import DB_CONFIG

def db_connect():
    """
    Create and return a connection to the MySQL database
    """
    try:
        connection = mysql.connector.connect(
            host=DB_CONFIG['host'],
            port=DB_CONFIG['port'],
            user=DB_CONFIG['user'],
            password=DB_CONFIG['password'],
            database=DB_CONFIG['database']
        )
        if connection.is_connected():
            print("Successfully connected to MySQL database")
            return connection
    except Error as e:
        print(f"Error while connecting to MySQL: {e}")
        raise

